Base64 encoding fix
Former-commit-id: 3d9936c7d6f80bebdf7795589eb8037d68d5df2e

+++ org.jkiss.dbeaver/core/src/org/jkiss/dbeaver/ui/export/wizard/DataExportJob.java
@@ -136,7 +136,7 @@ public class DataExportJob extends AbstractJob {
                 case BASE64:
                 {
                     writer.flush();
-                    Base64.encode(stream, (int) streamLength, writer);
+                    Base64.encode(stream, streamLength, writer);
                     break;
                 }
                 case HEX:
